+++ example/lib/generics_example.g.dart
@@ -6,29 +6,24 @@ part of 'generics_example.dart';
 // DataClassGenerator
 // **************************************************************************
 
-// ignore_for_file: annotate_overrides, unused_element
-
 mixin _$Response<T> {
   Response<T> get _self => this as Response<T>;
-
   Iterable<Object?> get _props sync* {
     yield _self.data;
   }
 
+  @override
   bool operator ==(Object other) =>
       identical(this, other) ||
       other is Response<T> &&
           runtimeType == other.runtimeType &&
           DataClass.$equals(_props, other._props);
-
+  @override
   int get hashCode => Object.hashAll(_props);
-
+  @override
   String toString() =>
       (ClassToString('Response', [T])..add('data', _self.data)).toString();
-
-  Response<T> copyWith({
-    T? data,
-  }) {
+  Response<T> copyWith({T? data}) {
     return Response(
       data: data ?? _self.data,
     );
@@ -36,17 +31,15 @@ mixin _$Response<T> {
 
   Response<T> change(void Function(_ResponseChanges<T> c) updates) =>
       (_ResponseChanges<T>._(_self)..update(updates)).build();
-
   _ResponseChanges<T> toChanges() => _ResponseChanges._(_self);
 }
 
 class _ResponseChanges<T> {
-  T data;
-
   _ResponseChanges._(Response<T> dc) : data = dc.data;
 
-  void update(void Function(_ResponseChanges<T> c) updates) => updates(this);
+  T data;
 
+  void update(void Function(_ResponseChanges<T> c) updates) => updates(this);
   Response<T> build() => Response(
         data: data,
       );
@@ -54,25 +47,24 @@ class _ResponseChanges<T> {
 
 mixin _$PaginatedResponse<T extends Object> {
   PaginatedResponse<T> get _self => this as PaginatedResponse<T>;
-
   Iterable<Object?> get _props sync* {
     yield _self.data;
     yield _self.total;
   }
 
+  @override
   bool operator ==(Object other) =>
       identical(this, other) ||
       other is PaginatedResponse<T> &&
           runtimeType == other.runtimeType &&
           DataClass.$equals(_props, other._props);
-
+  @override
   int get hashCode => Object.hashAll(_props);
-
+  @override
   String toString() => (ClassToString('PaginatedResponse', [T])
         ..add('data', _self.data)
         ..add('total', _self.total))
       .toString();
-
   PaginatedResponse<T> copyWith({
     T? data,
     int? total,
@@ -86,23 +78,25 @@ mixin _$PaginatedResponse<T extends Object> {
   PaginatedResponse<T> change(
           void Function(_PaginatedResponseChanges<T> c) updates) =>
       (_PaginatedResponseChanges<T>._(_self)..update(updates)).build();
-
   _PaginatedResponseChanges<T> toChanges() =>
       _PaginatedResponseChanges._(_self);
 }
 
 class _PaginatedResponseChanges<T extends Object>
     implements _ResponseChanges<T> {
-  T data;
-  int total;
-
   _PaginatedResponseChanges._(PaginatedResponse<T> dc)
       : data = dc.data,
         total = dc.total;
 
+  @override
+  T data;
+
+  int total;
+
+  @override
   void update(void Function(_PaginatedResponseChanges<T> c) updates) =>
       updates(this);
-
+  @override
   PaginatedResponse<T> build() => PaginatedResponse(
         data: data,
         total: total,
@@ -111,25 +105,22 @@ class _PaginatedResponseChanges<T extends Object>
 
 mixin _$ListResponse<T> {
   ListResponse<T> get _self => this as ListResponse<T>;
-
   Iterable<Object?> get _props sync* {
     yield _self.data;
   }
 
+  @override
   bool operator ==(Object other) =>
       identical(this, other) ||
       other is ListResponse<T> &&
           runtimeType == other.runtimeType &&
           DataClass.$equals(_props, other._props);
-
+  @override
   int get hashCode => Object.hashAll(_props);
-
+  @override
   String toString() =>
       (ClassToString('ListResponse', [T])..add('data', _self.data)).toString();
-
-  ListResponse<T> copyWith({
-    List<T>? data,
-  }) {
+  ListResponse<T> copyWith({List<T>? data}) {
     return ListResponse(
       data: data ?? _self.data,
     );
@@ -137,18 +128,19 @@ mixin _$ListResponse<T> {
 
   ListResponse<T> change(void Function(_ListResponseChanges<T> c) updates) =>
       (_ListResponseChanges<T>._(_self)..update(updates)).build();
-
   _ListResponseChanges<T> toChanges() => _ListResponseChanges._(_self);
 }
 
 class _ListResponseChanges<T> implements _ResponseChanges<List<T>> {
-  List<T> data;
-
   _ListResponseChanges._(ListResponse<T> dc) : data = dc.data;
 
+  @override
+  List<T> data;
+
+  @override
   void update(void Function(_ListResponseChanges<T> c) updates) =>
       updates(this);
-
+  @override
   ListResponse<T> build() => ListResponse(
         data: data,
       );