+++ example/lib/typedef_example.g.dart
@@ -6,11 +6,8 @@ part of 'typedef_example.dart';
 // DataClassGenerator
 // **************************************************************************
 
-// ignore_for_file: annotate_overrides, unused_element
-
 mixin _$Extra {
   Extra get _self => this as Extra;
-
   Iterable<Object?> get _props sync* {
     yield _self.jsonMap;
     yield _self.nullableJsonMap;
@@ -18,21 +15,21 @@ mixin _$Extra {
     yield _self.product2;
   }
 
+  @override
   bool operator ==(Object other) =>
       identical(this, other) ||
       other is Extra &&
           runtimeType == other.runtimeType &&
           DataClass.$equals(_props, other._props);
-
+  @override
   int get hashCode => Object.hashAll(_props);
-
+  @override
   String toString() => (ClassToString('Extra')
         ..add('jsonMap', _self.jsonMap)
         ..add('nullableJsonMap', _self.nullableJsonMap)
         ..add('product', _self.product)
         ..add('product2', _self.product2))
       .toString();
-
   Extra copyWith({
     JsonMap<int>? jsonMap,
     NullableJsonMap<double?>? nullableJsonMap,
@@ -49,24 +46,25 @@ mixin _$Extra {
 
   Extra change(void Function(_ExtraChanges c) updates) =>
       (_ExtraChanges._(_self)..update(updates)).build();
-
   _ExtraChanges toChanges() => _ExtraChanges._(_self);
 }
 
 class _ExtraChanges {
-  JsonMap<int>? jsonMap;
-  NullableJsonMap<double?> nullableJsonMap;
-  be.Product product;
-  be.Product product2;
-
   _ExtraChanges._(Extra dc)
       : jsonMap = dc.jsonMap,
         nullableJsonMap = dc.nullableJsonMap,
         product = dc.product,
         product2 = dc.product2;
 
-  void update(void Function(_ExtraChanges c) updates) => updates(this);
+  JsonMap<int>? jsonMap;
 
+  NullableJsonMap<double?> nullableJsonMap;
+
+  be.Product product;
+
+  be.Product product2;
+
+  void update(void Function(_ExtraChanges c) updates) => updates(this);
   Extra build() => Extra(
         jsonMap: jsonMap,
         nullableJsonMap: nullableJsonMap,