+++ example/test/basic_example_test.dart
@@ -4,18 +4,9 @@ import 'package:test/scaffolding.dart';
 
 void main() {
   group('basic_example', () {
-    final mouse1 = Product(
-      id: 1,
-      title: 'mouse',
-    );
-    final mouse2 = Product(
-      id: 1,
-      title: 'mouse',
-    );
-    final monitor = Product(
-      id: 2,
-      title: 'monitor',
-    );
+    final mouse1 = Product(1, 'mouse');
+    final mouse2 = Product(1, 'mouse');
+    final monitor = Product(2, 'monitor');
 
     test('Success data class comparison', () {
       expect(mouse1, mouse2);