+++ mek_data_class_generator/lib/src/writers/equality_writer.dart
@@ -1,3 +1,4 @@
+import 'package:code_builder/code_builder.dart';
 import 'package:mek_data_class_generator/src/specs.dart';
 import 'package:mek_data_class_generator/src/writers/writer.dart';
 
@@ -8,28 +9,28 @@ class EqualityWriter extends Writer {
   @override
   bool get available => classSpec.comparable;
 
+  @override
+  bool get needMixinMethodSelf => _fieldSpecs.isNotEmpty;
+
   late final _fieldSpecs = fieldSpecs.where((e) => e.isParam && e.comparable);
 
   @override
-  Iterable<String> writeMethods() sync* {
-    yield _generatePropsMethod();
-    yield _generateEqualMethod();
-    yield _generateHashcodeMethod();
+  Iterable<Method> creteMixinMethods() sync* {
+    yield _createMixinMethodProps();
+    yield _createMixinMethodEqual();
+    yield _createMixinMethodHashcode();
   }
 
-  String _generatePropsMethod() {
-    return '''
-  Iterable<Object?> get _props sync* {
-${_writePropsFields().join('\n')}
-  }''';
-  }
-
-  Iterable<String> _writePropsFields() sync* {
-    for (var field in _fieldSpecs) {
-      if (field.equality != null) continue;
-
-      yield '      yield _self.${field.name};';
-    }
+  Method _createMixinMethodProps() {
+    final props = _fieldSpecs.where((field) => field.equality == null).map((field) {
+      return 'yield _self.${field.name};';
+    });
+    return Method((b) => b
+      ..returns = Reference('Iterable<Object?>')
+      ..type = MethodType.getter
+      ..name = '_props'
+      ..modifier = MethodModifier.syncStar
+      ..body = Code(props.join()));
   }
 
   Iterable<String> _generateEquals() sync* {
@@ -40,19 +41,27 @@ ${_writePropsFields().join('\n')}
     }
   }
 
-  String _generateEqualMethod() {
+  Method _createMixinMethodEqual() {
     var equals = _generateEquals().join(' && ');
     if (equals.isNotEmpty) equals = ' && $equals';
 
-    return '''
-  bool operator ==(Object other) =>
-      identical(this, other) ||
+    final body = Code('''identical(this, other) ||
       other is ${classSpec.self.typedName} &&
           runtimeType == other.runtimeType &&
-          DataClass.\$equals(_props, other._props)$equals;''';
+          DataClass.\$equals(_props, other._props)$equals''');
+
+    return Method((b) => b
+      ..annotations.add(Annotations.override)
+      ..returns = Refs.bool
+      ..name = 'operator=='
+      ..requiredParameters.add(Parameter((b) => b
+        ..type = Refs.object
+        ..name = 'other'))
+      ..lambda = true
+      ..body = body);
   }
 
-  Iterable<String> _generateHashcodes() sync* {
+  Iterable<String> _codeEqualityHashcode() sync* {
     for (var field in _fieldSpecs) {
       if (field.equality == null) continue;
 
@@ -60,12 +69,17 @@ ${_writePropsFields().join('\n')}
     }
   }
 
-  String _generateHashcodeMethod() {
+  Method _createMixinMethodHashcode() {
     var propsStr = '_props';
-    final hascodes = _generateHashcodes().join(', ');
-    if (hascodes.isNotEmpty) propsStr += '.followedBy([$hascodes])';
+    final equalityHascode = _codeEqualityHashcode().join(', ');
+    if (equalityHascode.isNotEmpty) propsStr += '.followedBy([$equalityHascode])';
 
-    return '''
-    int get hashCode => Object.hashAll($propsStr);''';
+    return Method((b) => b
+      ..annotations.add(Annotations.override)
+      ..returns = Refs.int
+      ..type = MethodType.getter
+      ..name = 'hashCode'
+      ..lambda = true
+      ..body = Code('Object.hashAll($propsStr)'));
   }
 }